# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "endpoints",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/houses": {
      "get": {
        "tags": [
          "house"
        ],
        "operationId": "get_houses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "last_evaluated_key",
            "in": "query",
            "description": "Last evaluated key from previous response",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_HouseResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/maps": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "get_map",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/maps/requirements": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "post_requirement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "HouseResponse": {
        "type": "object",
        "required": [
          "h3_index",
          "address",
          "url",
          "lat",
          "lon"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "h3_index": {
            "type": "string"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "Location": {
        "type": "object",
        "required": [
          "id",
          "address",
          "h3_index",
          "lat",
          "lng"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "h3_index": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "MapRequest": {
        "type": "object",
        "required": [
          "city_code",
          "requirement_ids"
        ],
        "properties": {
          "city_code": {
            "type": "string"
          },
          "requirement_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "MapResponse": {
        "type": "object",
        "required": [
          "tiles"
        ],
        "properties": {
          "tiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapTileResponse"
            }
          }
        }
      },
      "MapTileResponse": {
        "type": "object",
        "required": [
          "h3_index",
          "mean_score",
          "requirement_scores"
        ],
        "properties": {
          "h3_index": {
            "type": "string"
          },
          "mean_score": {
            "type": "integer",
            "format": "int32"
          },
          "requirement_scores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequirementScoreResponse"
            }
          }
        }
      },
      "PaginatedResponse_HouseResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "h3_index",
                "address",
                "url",
                "lat",
                "lon"
              ],
              "properties": {
                "address": {
                  "type": "string"
                },
                "h3_index": {
                  "type": "string"
                },
                "lat": {
                  "type": "number",
                  "format": "double"
                },
                "lon": {
                  "type": "number",
                  "format": "double"
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "last_evaluated_key": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "RequirementRequest": {
        "type": "object",
        "required": [
          "requirement_id",
          "city_code",
          "travel_mode",
          "locations",
          "tolerated_duration"
        ],
        "properties": {
          "city_code": {
            "type": "string"
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Location"
            }
          },
          "requirement_id": {
            "type": "string",
            "format": "uuid"
          },
          "tolerated_duration": {
            "type": "integer",
            "format": "int32"
          },
          "travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          }
        }
      },
      "RequirementResponse": {
        "type": "object"
      },
      "RequirementScoreResponse": {
        "type": "object",
        "required": [
          "requirement_id",
          "score"
        ],
        "properties": {
          "requirement_id": {
            "type": "string",
            "format": "uuid"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TravelMode": {
        "type": "string",
        "enum": [
          "Driving",
          "Walking",
          "Bicycling",
          "PublicTransport"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "house",
      "description": "House endpoints"
    },
    {
      "name": "map",
      "description": "Map endpoints"
    }
  ]
}